Fixed  Issue 462: 	USER 130 crash when shouting at an empty shoutbox
ShoutAtTextUserLC was being called, which looks up a iList[] item, and causes an array out-of-bounds panic.

I also noticed if your usename is all lower case (e.g. "myusername") in Mobbler, but mixed case on Last.fm (e.g. "MyUserName"), then if there's a single shout from yourself, it shows it twice as: 
Shout as myusername
Shout as MyUserName
But it should show it only once. So I've fixed that too (without making a new bug, is that ok?).

git-svn-id: https://mobbler.googlecode.com/svn/trunk@715 3adbc9d6-a949-0410-9cc0-931bda5f43ff