Fixed Issue 468: 	KERN-EXEC 3 when playing tag radio
OK, so that tag exists and has been used once, and there are no tracks. 
http://www.last.fm/tag/muay%20thai

The KERN-EXEC is coming from CMobblerParser::ParseRadioTuneL(aXml). aXml contains:

===
<?xml version="1.0" encoding="utf-8"?>
<lfm status="failed">
    <error code="20">There is not enough content to play this station</error></lfm>
===

And this is the problem line because there are no <station> or <name> tags:

	CMobblerString* station(CMobblerString::NewL(domFragment->AsElement().Element
(_L8("station"))->Element(_L8("name"))->Content()));

A normal response would look like:

===
<?xml version="1.0" encoding="utf-8"?>
<lfm status="ok">
<station>
    <type>globaltags</type>
    <name>blues Tag Radio</name>
    <url>http://www.last.fm/listen/globaltags/blues</url>
    <supportsdiscovery>1</supportsdiscovery>
</station></lfm>
===

So I've fixed this by checking the <lfm status="XXX">. If "ok", do as before. Otherwise grab the contents of the <error> tag as the station name and carry on as before: Last.fm also sends the last working playlist (e.g. blues tag) and so Mobbler just carries on playing that, but with the error in the navipane.

git-svn-id: https://mobbler.googlecode.com/svn/trunk@736 3adbc9d6-a949-0410-9cc0-931bda5f43ff