Re-fixed Issue 485: 	Mobbler cannot be closed from the task manager
Allow Mobbler to keep running when pressing the red end key, but exit it when pressing C from the task manager.

Change the EEikCmdExit to actually exit as before, but also add HandleWsEventL() and do nothing if the event type is KAknUidValueEndKeyCloseEvent (i.e. red end key to minimise). Otherwise handle it as normal.