Fix for issue 562: Visit Last.fm is in the Friends' View list. Just moved the commands for opening the webpage to the online section in the .hrh, so they're not added to the view submenu.
This means if you try and visit a web page from the status view, it will also demand the username and password are present, when strictly speaking you don't need them.

That's not a huge problem, and we should just get rid of that check anyway and make sure there's always a username/password. We'll fix that properly with a wizard in issue 563.