Implemented Issue 408: 	Fetch song lyrics, ready for review.
This writes the lyrics to a text file and opens it in the text viewer, a nice and simple UI.

At the moment it's using a weekly user ID key below for testing with Lyricsfly, which returns the first 30% of lyrics, if it doesn't work make sure you have the latest from http://www.lyricsfly.com/api/#doc

I'll apply for a permanent user ID from Lyricsfly.

Once the biography viewer thingy is ready, we could look to showing the lyrics with that.