Finishing off Issue 408: 	"Fetch song lyrics".
I received the permanent user ID key from Lyricsfly. They ask us to protect the key, so this shouldn't be distributed. It's included from a seperate header file if the PERMANENT_LYRICSFLY_ID_KEY macro is defined, otherwise a testing key is used.

Should the macro be on or off by default in the source?

Furthermore, as requested by Lyricsfly to improve their data, it also adds a link back to Lyricsfly to allow corrections at the very end: "Make corrections:\r\nhttp://lyricsfly.com/search/correction.php?[CHECKSUM]&id=[SONG ID]". I've not localised it because all lyrics text ends with an unlocalised "Lyrics delivered by lyricsfly.com", but we can easily localise it, let me know if we should.

And it's ok to add the hyperlink to the text file, because the text viewer's "Automatic find" thingy can conveniently hyperlink the URL. Clicking it opens the webpage in the browser, and it's quite lightweight and seems to work ok on my phone.