A few things.
Updates for issue 634.
Refined the warning text.
Generalise and re-use the CMobblerSystemCloseGlobalQuery (renamed CMobblerGlobalQuery) for use by the warning as well.
Note: iStatus contains the pressed softkey value and not an error code -- OK is -2 -- so stop the CActiveScheduler in RunL() regardless of a positive value.
I also changed the query to a warning for both system close and old scrobble warnings.
Only show the warning once per day, and show it whenever the status view is activated (e.g. on launch, or after returning from settings), and also when scrobbling a new track (in case Mobbler is always running).

Add limit=1 to get only one event when checking if this band is on tour. This will be live on Last.fm in a couple of weeks.

Only initialise a load of non-essential and possibly rarely used member variables when needed instead of in CMobberAppUi::ConstructL(). e.g. iAlarmTimer, iSleepTimer, iBrowserLauncher, iMobblerDownload, iLocation

Removed some duplicate code from mobblertrack.cpp, no point trying to find %album%.jpg/gif/png twice.

Minor code review bits for the local events thingy.