Update following code review on r8818df573a by gw111zz:
"This was actually in the original code I wrote but I've just seen it now. It looks like artist, artistInfo, imageUrl, tagsText and similarArtistsText aren't needed any more after this line since the content have been copied into artistInfoHtml. So all of the former HBufC8s can be destroyed now rather than existing for the duration of time the user is viewing the biography. This should save memory since the browser control and the HTML biography content take up quite a lot as it is."