Fix for Issue 733:	Update our usage of auth.getMobileSession === Using authToken with auth.getMobileSession has been deprecated. We should move to sending the plaintext password instead.
Docs: http://www.last.fm/api/show/auth.getMobileSession

Note that it also must now be sent as an https post.
===

In CMobblerTransaction::SubmitL(), check for iConnection.iWebServicesSessionKey
before trying to add it as a field because it won't be there for
auth.getMobileSession and would otherwise cause a KERN-EXEC 3.

Also removed unused CMobblerLastFmConnection::CreateAuthTokenL().

Should we use https for everything or just auth.getMobileSession?